Add New in Pike note to using db_sync check
This change adds a note (*New in Pike*) to the section for running
db_sync --check to clarify which release this feature was added. This
is similar to other sections in this document that specify the release
that a feature was added in and when it is available.

Change-Id: I7d778fa249997bb24e4563d6960910b93637bbf2
Closes-Bug: #1734549
(cherry picked from commit 756d28149043f298ba25c654e1ef31d7e690c909)