Fix role schema in trust object
Previously, we weren't doing any validation on the roles attribute of a
trust except to validate that it was an array. A hasty glance, however,
would lead you to believe that it was validating an array of
parameter_types.id_string[1] and so we translated that to the new role
object validation. However, id_string doesn't include some valid role
names like _member_. This patch updates the role name schema to match
parameter_types.name, which is the same as the schema for the main role
object.

[1] http://git.openstack.org/cgit/openstack/keystone/tree/keystone/trust/schema.py?id=62f9e57cd81dc98c5816da9fa483d385b4c1a66c#n41

Change-Id: I83aafc7a96e81a9b6b1056b39cd8c5d23676c014
Closes-bug: #1734871