Remove private methods for v2.0 and v3 tokens
The fernet token formatter had a couple of private methods in order
to understand how to pull data from token references depending on if
it was a v2.0 token or a v3 token. This isn't needed anymore because
the v2.0 token API were removed in queens.

This commit simplifies that logic and removes the extra case.

Change-Id: I18f1f87aa4af7fbaf66767a016bb0e15f7384f05