Add scope_types to service policies
Services in keystone are entities that are specific to the deployment
system. It makes sense that you'd need a system-scoped token in order
to manage services.

This commit adds `system` as the scope type for services. Now when
project scoped tokens are used to manage services, a warning will
be logged. Operators will also have the ability to turn on the
`enforce_scope` option via oslo.policy to require system-scoped
tokens to interact with services.

Change-Id: If3bd1dd5680e397da16a493d32326207ada05404