Add scope_types to grant policies
Since grants can exist on either the system or a project, it makes
sense to have them associated to both `system` and `project` scopes.
But until we have logic in keystone that can enforce those checks, we
keep the scope_types set to `system`. This ensures backwards
compatibility and helps resolve admin-ness issues until we get to
that point.

Change-Id: I310c88bae7986e3163408f621e266f96fd255edc