Add scope_types to identity provider policies
I set the policies for these to be `system` even though it makes
sense to have them be `system` and `project` scopes once it is
possible to add federated identity without having to roll apache
configuration files.

From now on, a warning will be logged when a project-scoped token is
used to manage identity providers. Operator will have the option to
enforce scope checking if they enable `enforce_scope` via
oslo.policy's configuration.

Change-Id: I2aa2ff866e1622d8dc2ce813956ac869be792b5f