Document scope_types for credential policies
This change adds scope types for all credential policies and sets it
to `system` and `project`, but leaves them commented out. When
keystone supports the ability to enforce credential checks in code
based on the system-scoped tokens versus project-scoped tokens, we
can uncomment the scope_types and allow them to be enforced.

This commit describes how those API should behave in different
circumstances and gives us something track using bugs.

Change-Id: I61b15f3b2865e516a5358b9caacdb7ca8f6f1d17