Improve exception logging with 500 response
Currently when keystone throws a 500 Error, depending on the actual
exception type, it can log the message as an exception or as a warning.

Specifically, if the server throws an exception.UnexpectedError, it
does not log this as an exception; it simply logs it as a warning. This
patch set logs the error as an exception if exception.Error is an
exception.UnexpectedError.

Change-Id: Ia47cc11378ec64d59b7403cb8a284c764148d7a9
Co-Authored-By: Tin Lam <tin@irrational.io>
Closes-Bug: #1717962