Handle InvalidScope exception from oslo.policy
oslo.policy has the ability to check the scope of the request
against the scope of the operation and it will throw an InvalidScope
exception if they don't match. If we catch this exception, we should
throw a Forbidden instead because the caller doesn't have the
necessary role on the intended scope to perform the action.

Change-Id: I8805db48b2ce82397ce70cdb4d6b6d8178603339