Deprecate [trust]/enabled option
It's currently possible to disable the trusts API and trusts are
therefore essentially an API extension and not a first-class member of
the keystone API. By now, trusts are a widely-used feature and are
essential for services like Heat. Allowing optional APIs is an
interoperability nightmare that we should discourage. This patch marks
the config flag to enable trusts as deprecated so that in the future it
will not be possible to disable it.

bp deprecated-as-of-queens

Change-Id: Iac454ddf7878b8f2705e7abb44181f1a19af015f