Fix federation unit test
A patch[1] was introduced to demonstrate a bug in federation token
auth[2]. The bug was later fixed[3] but one of the new tests was never
un-WIP'd. It continued to "fail as expected" because it was using a
method that didn't exist, not failing due to the bug it was supposed to
expose. This patch fixes the test and un-WIPs it.

[1] https://review.openstack.org/#/c/229125
[2] https://bugs.launchpad.net/keystone/+bug/1501032
[3] https://review.openstack.org/#/c/431181/

Change-Id: I4d76362256e41fafc87d413cde090fb12450ec83