Add system column to app cred table
While the application credential logic was going through review we
noticed the project_id column of the table was not nullable. It is
possible for this to hinder the ability to create system-level
application credentials in the future.

For now, let's create a system column for application credentials
and alter the table so that project_ids can be nullable. This will
make it easier to expand application credential usage in the future.

bp application-credentials

Change-Id: I4e1104f95cc3c7567ee8f6edfe8515d45d154a9f