Update OBS install docs for v2 removal
With the removal of the v2 API, there is no reason to keep listening on
multiple ports. Update the OBS install guide to only mention the one
port. The openSUSE openstack-keystone package does not provide a default
vhost config file so we can update this independently of any package
changes in the distro. This also removes a few incorrect notes, one
claiming that the distro package installed and started the nonexistent
eventlet service and one claiming that port 5000 only allowed non-admin
access.

Change-Id: Ic06af94335598e0aadac20874d177e531069548a