Expose bug in system assignment when deleting users
Project and domain role assignment are cleaned up when deleting
users. This commit introduces a test case that shows this isn't the
case for system role assignments. A subsequent patch will implement
a fix to make sure system role assignments are removed when users
are deleted, to be consistent with project and domain assignments.

Change-Id: I1a1e7395f462159037e939aa143e9e24aefb1841
Partial-Bug: 1749264