Make assertValidFernetKey assertion more robust
This assertion was a bit misleading in that it seemed to validate
one key at a time. But the implementation of the assertion actually
assumes it is given a list of string values. This commit fixes the
name of the assertion to be more clear and it also makes the
assertion iterate all values given, which ensures each key in the
list is valid, not just the first one.

Change-Id: Ifc9542e4e7f57f54d473183c97f616f92cb0961b