Remove unnecessary config overrides from fernet tests
The fernet token unit tests had a bunch of config_fixture overrides to
populate the fernet max_active_keys configuration option with its
default value.  Since it is just setting these to the default, it's
not necessary to provide overrides when it's already established in
keystone.tests.unit.core.TestCase.

Change-Id: Id09459104d9b5ec412ec60211ef71b7c7c62b14f