Add release note for keystone v2 removal
It is possible that some users might be slightly affected by this
change. Add a release note about the removal of keystone v2 in queens.

Since Queens already went out the door, this change is being proposed
only to the stable/queens branch so that it is properly rendered in the
Queens release notes page.

Change-Id: I9aaf7c9eff9000401bca9d4884675f0b1d9aeb3e