Add logging for xmlsec1 installation
Keystone uses a library called xmlsec1 to create SAML assertions when
acting as an identity provider. If this library isn't present and
someone attempts to authenticate, keystone will throw an HTTP 500.
The only thing the error says is that a file or directory doesn't
exist.

This patch uses subprocess to check if the provided binary actually
exists on the system and handles cases when it isn't and logs a
useful message for operators.

Change-Id: I41cf87702df5389c1424d35f0abcef9c16301450
Closes-Bug: 1750917