Add note to keystone-manage bootstrap doc
A lot of people are very surprised that we no longer reference port
35357 in our Ubuntu (and SUSE) install guides. Add a note to clarify
that this is not a bug and we do really mean it.

This does not change the RDO install guide because our guide still
instructs users to use port 35357 because the RDO package still includes
an Apache vhost file that uses 35357.

Change-Id: I334ba888190705a345d50cebe577b832753f202c
Related-bug: #1755026
Related-bug: #1755511
Related-bug: #1756178