Remove references to v2.0 from external developer doc
The external developer document we have attempts to clarify various
concepts in keystone and make it easier for other developers writing
other services. Now that we've removed the v2.0 API, it makes sense
to update this documentation to refer to v2.0 in the past tense. Some
parts of the document seemed specific to operator documentation, which
has been either removed or reworked to the intended operators, other
service developers.

Change-Id: I809150f8b77a813e2300760fdcb1d11cfa8ca732