Removal of deprecated direct driver loading
The direct loading of drivers was deprecated in Liberty and noted
to be removed in Newton. This patch cleans up the deprecation and
fixes the unit tests. Some of the example request/response json
message bodies of the domain config API were also updated to the
correct way of loading a driver now.

bp removed-as-of-rocky

Change-Id: If3f4c2303da6e264e5e0d73280cc21fa01a3cfd4