Log warning when using token_flush
The `keystone-manage token_flush` command was useful when we
supported token format that were actually written to disk. Now that
keystone has moved towards formats that don't require physical
storage in the name of scalability, we no longer need the
token_flush utility.

This commit logs a warning when using `keystone-manage token_flush`
so that people are aware they don't need to use it anymore if they
are using supported upstream non-persistent token providers.

Change-Id: I14d1e8a3ae367c8f13b6819a1a216f3f8f6b001a
Closes-Bug: 1759289