Introduce new TokenModel object
Our previous token model (``KeystoneToken``) was built to accept
complete token responses and use reflection to be more pythonic. Now
that we're no longer supporting token providers or storage layers
that rely on passing around full token responses, we can introduce
a more pythonic way of building tokens.

This commit introduces a new TokenModel object that is intended to be
used with composition, starting with a blank state and allowing parts
of the application to fill in the blanks. This should hopefully make
our entire authentication and validation API paths more clear and
understandable.

The following are a few benefits of adopting the new model.

First, this helps us simplify the interface between the token provider
API and the actual token providers. The token providers shouldn't have
to understand the API contract in order to work. Likewise, if we add
another API version, we shouldn't need to change multiple token
providers to ensure the API contract isn't violated depending on which
one is being used.

Second, this forces us to translate the token model into a token
response in a single place (e.g. the controller), isolating where we
represent API contracts. The is mostly an extension of the first
point, but it helps us be explicit in what we build and give to a
user. This type of architecture should help us avoid letting bugs like
bug 1763510 slip through our fingers.

Third, and somewhat subjectively, this makes the code easier to read
by dealing with objects in pythonic ways as opposed to understanding
large data structures in the controller, manager, and driver layers of
a subsystem.

Partial-Bug: 1778945
Change-Id: I2e461e7fdf4f47547e07fc279ec4e6fbd86f23ad