Update IdP sql model
Base on the database schema, the domain_id column in identity_provider
is not unique and has the ForeignKey for project.id. But the IdP sql
model is different. It marks the domain_id is unique and the ForeignKey
is lost.

This patch removes the unique restriction and adds the FK back, ultimately
making the relationship between domains and identity provider 1:many.

Change-Id: I13ecb0ab0434f5614f31d151e708f299cf8e8adb
Partial-bug: #1760843