Use the provider_api module in limit controller
Last release we spent a bunch of time going through and updating the
code base to use the keystone/common/provider_api.py module. It
looks like the unified limit controller was missed through.

This commit updates the unified limit controller to be consistent
with the rest of the keystone in not using `self` to access other
subsystems in keystone.

Change-Id: I86fa5a6e889d24779b2451b8402d0e7bd92a64ba