Implement enforcement model logic in Manager
This commit adds the necessary bits to return the enforcement model
used by a deployment in the limit Manager. A subsequent patch will
expose this functionality via the API.

This commit also starts introdcuing the concept of enforcement
models and puts the flat model into its own module. This will make
more sense and be easier to maintain as we implement more models.

Change-Id: I32227eb0023e6b6ce699909fabb60a63a07f0969
Related-Bug: 1765193
Related-Bug: 1768572