Correct test_v3_oauth1.test_bad_authorizing_roles_name
`test_bad_authorizing_roles_name` was incorrecting using .admin_request
instead of self.put. During the conversion to flask, this results in the
incorrect HTTP 404 (server text/html instead of application
(application/json); the incorrect behavior was due to dispatching to
and un-routed URI. In the legacy keystone wsgi, this was still handled
by the application and erroneously raised a 404 that looked correct.

The use of .put instead of .admin_request simply adds in the /v3 prefix
to the url returned by .authorize_request_token making the routing
now hit the proper controller method.

Change-Id: I1bdfbcaf1c3b6d0d86e2819b65d73af45bc7644d