Correct test_v3_oauth1.test_change_user_password_also_deletes_tokens
The `test_change_user_password_also_deletes_tokens` test was not
utilizing .get and therefore was not getting the '/v3' prepended to
the /auth/tokens request path. This resulted in an erroneous 404
response from the server (in flask this generates as an text/html
404 instead of an application handled application/json 404). This
test would always succeed in the last step as the token was not
actually validated. The 404 response was generated due to the
lack of routing instead of the lack of a valid token.

Change-Id: I850efdbc20281e454828e40664b58a28036f79dd